Add files via upload
I began making the solution and name for the project to Lab01. After that I had to start inside the Bracket to start the Console. Starting with Console I needed proper capitalization starting with new words, known as Pascal Case. Writing Console.WriteLine it displays the argument in the console window. Console.ReadLine tells the window to display what I wrote for the Console.WriteLine. I needed to add parenthesis and double quotation marks to have the command inside showing what I'm going to say in my window display. Also the Semi colon to finish off the line. Inside the parenthesis I had whats called Blank Lines so it's easier for the apps to read. The Input directs the user to take the specific action. The IDE that was used applied syntax-color highlightening for what the words will do, and show if there is an error.